requirements_cuda11.txt
ipdb
matplotlib
seaborn
pandas
tqdm
jupyter
equinox==0.10.4
optax==0.1.7
einops
scikit-learn
tensorflow
tensorflow-datasets
hydra-core
hydra-submitit-launcher
wandb
jax==0.4.7
chex==0.1.7
jaxtyping==0.2.15
scipy==1.10.1
